[AIRFLOW-5959] AIP-21 Move contrib/*/jira to providers
