+++ docs/conf.py
@@ -239,6 +239,7 @@ exclude_patterns = [
     '_api/base_serialization/index.rst',
     '_api/serialized_baseoperator/index.rst',
     '_api/serialized_dag/index.rst',
+    '_api/airflow/providers/jira',
     'autoapi_templates',
     'howto/operator/gcp/_partials',
 ]
