+++ tests/providers/jira/hooks/test_jira.py
@@ -21,8 +21,8 @@
 import unittest
 from unittest.mock import Mock, patch
 
-from airflow.contrib.hooks.jira_hook import JiraHook
 from airflow.models import Connection
+from airflow.providers.jira.hooks.jira import JiraHook
 from airflow.utils import db
 
 jira_client_mock = Mock(
@@ -38,7 +38,7 @@ class TestJiraHook(unittest.TestCase):
                 host='https://localhost/jira/', port=443,
                 extra='{"verify": "False", "project": "AIRFLOW"}'))
 
-    @patch("airflow.contrib.hooks.jira_hook.JIRA", autospec=True,
+    @patch("airflow.providers.jira.hooks.jira.JIRA", autospec=True,
            return_value=jira_client_mock)
     def test_jira_client_connection(self, jira_mock):
         jira_hook = JiraHook()
