+++ tests/providers/jira/sensors/test_jira.py
@@ -22,8 +22,8 @@ import unittest
 from unittest.mock import Mock, patch
 
 from airflow import DAG
-from airflow.contrib.sensors.jira_sensor import JiraTicketSensor
 from airflow.models import Connection
+from airflow.providers.jira.sensors.jira import JiraTicketSensor
 from airflow.utils import db, timezone
 
 DEFAULT_DATE = timezone.datetime(2017, 1, 1)
@@ -59,7 +59,7 @@ class TestJiraSensor(unittest.TestCase):
                 host='https://localhost/jira/', port=443,
                 extra='{"verify": "False", "project": "AIRFLOW"}'))
 
-    @patch("airflow.contrib.hooks.jira_hook.JIRA",
+    @patch("airflow.providers.jira.hooks.jira.JIRA",
            autospec=True, return_value=jira_client_mock)
     def test_issue_label_set(self, jira_mock):
         jira_mock.return_value.issue.return_value = minimal_test_ticket
